+++ src/main/java/org/openlmis/referencedata/domain/measurement/TemperatureMeasurement.java
@@ -1,0 +1,48 @@
+/*
+ * This program is part of the OpenLMIS logistics management information system platform software.
+ * Copyright © 2017 VillageReach
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms
+ * of the GNU Affero General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License for more details. You should have received a copy of
+ * the GNU Affero General Public License along with this program. If not, see
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
+ */
+
+package org.openlmis.referencedata.domain.measurement;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import javax.persistence.Embeddable;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Embeddable
+@NoArgsConstructor
+public class TemperatureMeasurement extends BaseMeasurement {
+
+  private String temperatureMeasurementUnitCode;
+
+  @Override
+  public List<String> getCodeListVersion() {
+    List<String> list = Stream.of(TemperatureUnitCode.values())
+            .map(TemperatureUnitCode::name)
+            .collect(Collectors.toList());
+    return list;
+  }
+
+  public TemperatureMeasurement(Double value, String temperatureMeasurementUnitCode) {
+    super(value);
+    this.temperatureMeasurementUnitCode = temperatureMeasurementUnitCode;
+  }
+
+}