+++ src/main/java/org/openlmis/referencedata/validate/OrderableValidator.java
@@ -21,7 +21,10 @@ import static org.openlmis.referencedata.util.messagekeys.OrderableMessageKeys.E
 import static org.openlmis.referencedata.util.messagekeys.OrderableMessageKeys.ERROR_PRODUCT_CODE_REQUIRED;
 import static org.openlmis.referencedata.util.messagekeys.OrderableMessageKeys.ERROR_ROUND_TO_ZERO_REQUIRED;
 
+import org.openlmis.referencedata.domain.measurement.TemperatureMeasurement;
+import org.openlmis.referencedata.domain.measurement.VolumeMeasurement;
 import org.openlmis.referencedata.dto.OrderableDto;
+import org.openlmis.referencedata.util.messagekeys.OrderableMessageKeys;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.ValidationUtils;
@@ -32,6 +35,10 @@ import org.springframework.validation.Validator;
  */
 @Component
 public class OrderableValidator implements BaseValidator {
+
+  static final String MINIMUM_TOLERANCE_TEMPERATURE = "minimumToleranceTemperature";
+  static final String MAXIMUM_TOLERANCE_TEMPERATURE = "maximumToleranceTemperature";
+  static final String IN_BOX_CUBE_DIMENSION = "inBoxCubeDimension";
 
   /**
    * Checks if the given class definition is supported.
@@ -61,5 +68,91 @@ public class OrderableValidator implements BaseValidator {
     rejectIfNull(errors, "packRoundingThreshold", ERROR_PACK_ROUNDING_THRESHOLD_REQUIRED);
     rejectIfNull(errors, "roundToZero", ERROR_ROUND_TO_ZERO_REQUIRED);
     rejectIfNull(errors, "netContent", ERROR_NET_CONTENT_REQUIRED);
+
+    if (!errors.hasErrors()) {
+      OrderableDto dto = (OrderableDto) target;
+      validateToleranceTemperature(dto, errors);
+      validateVolumeMeasurement(dto, errors);
+    }
+
+  }
+
+  private void validateToleranceTemperature(OrderableDto dto, Errors errors) {
+    if (isNotToleranceTemperatureCodeSupported(dto.getMinimumToleranceTemperature())) {
+      rejectValue(errors, MINIMUM_TOLERANCE_TEMPERATURE, OrderableMessageKeys
+              .ERROR_MINIMUM_TOLERANCE_TEMPERATURE_UNIT_CODE_NOT_SUPPORTED);
+    }
+    if (isNotToleranceTemperatureCodeSupported(dto.getMaximumToleranceTemperature())) {
+      rejectValue(errors, MAXIMUM_TOLERANCE_TEMPERATURE, OrderableMessageKeys
+              .ERROR_MAXIMUM_TOLERANCE_TEMPERATURE_UNIT_CODE_NOT_SUPPORTED);
+    }
+    if (isMinTemperatureValueGreaterThanMaxTemperatureValue(
+            dto.getMinimumToleranceTemperature(),
+            dto.getMaximumToleranceTemperature())) {
+      rejectValue(errors, MINIMUM_TOLERANCE_TEMPERATURE, OrderableMessageKeys
+              .ERROR_MINIMUM_TOLERANCE_TEMPERATURE_VALUE);
+    }
+    if (isNotGivenToleranceTemperatureCode(dto.getMinimumToleranceTemperature())) {
+      rejectValue(errors, MINIMUM_TOLERANCE_TEMPERATURE, OrderableMessageKeys
+              .ERROR_MINIMUM_TOLERANCE_TEMPERATURE_UNIT_CODE_REQUIRED);
+    }
+
+    if (isNotGivenToleranceTemperatureCode(dto.getMaximumToleranceTemperature())) {
+      rejectValue(errors, MAXIMUM_TOLERANCE_TEMPERATURE, OrderableMessageKeys
+              .ERROR_MAXIMUM_TOLERANCE_TEMPERATURE_UNIT_CODE_REQUIRED);
+    }
+  }
+
+  private void validateVolumeMeasurement(OrderableDto dto, Errors errors) {
+    if (dto.getInBoxCubeDimension().getMeasurementUnitCode() != null
+            && dto.getInBoxCubeDimension().getValue() != null) {
+      if (isNotMeasurementUnitCodeSupported(dto.getInBoxCubeDimension())) {
+        rejectValue(errors, IN_BOX_CUBE_DIMENSION, OrderableMessageKeys
+                .ERROR_IN_BOX_CUBE_DIMENSION_UNIT_CODE_NOT_SUPPORTED);
+      }
+      if (!isMeasurementVolumeValuePositive(dto.getInBoxCubeDimension().getValue())) {
+        rejectValue(errors, IN_BOX_CUBE_DIMENSION, OrderableMessageKeys
+                .ERROR_IN_BOX_CUBE_DIMENSION_VALUE);
+      }
+    }
+    if (dto.getInBoxCubeDimension().getMeasurementUnitCode() == null
+            && dto.getInBoxCubeDimension().getValue() != null) {
+      rejectValue(errors, IN_BOX_CUBE_DIMENSION, OrderableMessageKeys
+              .ERROR_IN_BOX_CUBE_DIMENSION_UNIT_CODE_REQUIRED);
+    }
+  }
+
+  private boolean isNotToleranceTemperatureCodeSupported(
+          TemperatureMeasurement temperatureMeasurement) {
+    return isGivenToleranceTemperature(temperatureMeasurement)
+            && !temperatureMeasurement.getCodeListVersion()
+            .contains(temperatureMeasurement.getTemperatureMeasurementUnitCode());
+  }
+
+  private boolean isNotMeasurementUnitCodeSupported(VolumeMeasurement volumeMeasurement) {
+    return !volumeMeasurement.getCodeListVersion()
+            .contains(volumeMeasurement.getMeasurementUnitCode());
+  }
+
+  private boolean isMinTemperatureValueGreaterThanMaxTemperatureValue(
+          TemperatureMeasurement minTemperature, TemperatureMeasurement maxTemperature) {
+    return isGivenToleranceTemperature(minTemperature)
+            && isGivenToleranceTemperature(maxTemperature)
+            && minTemperature.getValue() > maxTemperature.getValue();
+  }
+
+  private boolean isMeasurementVolumeValuePositive(Double value) {
+    return value >= 0;
+  }
+
+  private boolean isGivenToleranceTemperature(TemperatureMeasurement temperatureMeasurement) {
+    return temperatureMeasurement.getTemperatureMeasurementUnitCode() != null
+            && temperatureMeasurement.getValue() != null;
+  }
+
+  private boolean isNotGivenToleranceTemperatureCode(
+          TemperatureMeasurement temperatureMeasurement) {
+    return temperatureMeasurement.getTemperatureMeasurementUnitCode() == null
+            && temperatureMeasurement.getValue() != null;
   }
 }